drivers: unify mbedtls_platform_entropy_poll and mbedtls_hardware_poll
Since the 2 functions are one the alternative of the other depending on
MBEDTLS_PLATFORM_GET_ENTROPY_ALT, let's unify their naming so that we
can clarify code a bit.
