+++ tests/suites/test_suite_psa_crypto_init.function
@@ -25,14 +25,8 @@ exit:
     return 0;
 }
 
-#if defined(MBEDTLS_PLATFORM_GET_ENTROPY_ALT)
-#define ENTROPY_POLL_MIN    MBEDTLS_ENTROPY_MIN_HARDWARE
-#else /* MBEDTLS_PLATFORM_GET_ENTROPY_ALT */
-#define ENTROPY_POLL_MIN    MBEDTLS_ENTROPY_MIN_PLATFORM
-#endif /* MBEDTLS_PLATFORM_GET_ENTROPY_ALT */
-
 #define ENTROPY_MIN_NV_SEED_SIZE                                        \
-    MAX(ENTROPY_POLL_MIN, MBEDTLS_ENTROPY_BLOCK_SIZE)
+    MAX(MBEDTLS_ENTROPY_POLL_PLATFORM_MIN, MBEDTLS_ENTROPY_BLOCK_SIZE)
 
 #include "psa_crypto_random_impl.h"
 #if defined(MBEDTLS_PSA_HMAC_DRBG_MD_TYPE)
@@ -100,19 +94,11 @@ static void custom_entropy_init(mbedtls_entropy_context *ctx)
     ctx->accumulator_started = 0;
     mbedtls_md_init(&ctx->accumulator);
 
-#if defined(MBEDTLS_PLATFORM_GET_ENTROPY_ALT)
     if (custom_entropy_sources_mask & ENTROPY_SOURCE_HARDWARE) {
-        mbedtls_entropy_add_source(ctx, mbedtls_hardware_poll, NULL,
-                                   MBEDTLS_ENTROPY_MIN_HARDWARE,
+        mbedtls_entropy_add_source(ctx, mbedtls_entropy_poll_platform, NULL,
+                                   MBEDTLS_ENTROPY_POLL_PLATFORM_MIN,
                                    MBEDTLS_ENTROPY_SOURCE_STRONG);
     }
-#else /* MBEDTLS_PLATFORM_GET_ENTROPY_ALT */
-    if (custom_entropy_sources_mask & ENTROPY_SOURCE_PLATFORM) {
-        mbedtls_entropy_add_source(ctx, mbedtls_platform_entropy_poll, NULL,
-                                   MBEDTLS_ENTROPY_MIN_PLATFORM,
-                                   MBEDTLS_ENTROPY_SOURCE_STRONG);
-    }
-#endif /* MBEDTLS_PLATFORM_GET_ENTROPY_ALT */
 #if defined(MBEDTLS_ENTROPY_NV_SEED)
     if (custom_entropy_sources_mask & ENTROPY_SOURCE_NV_SEED) {
         mbedtls_entropy_add_source(ctx, mbedtls_nv_seed_poll, NULL,
